+++ src/Docks/DocksOutput.twee
@@ -1,9 +1,9 @@
 :: DocksOutput
 \<<set $building to lib.findInArray($town.buildings, "key", $currentPassage.key)>>
-\<<set $currentPassage.availableRelationships to setup.docks.get.customers>>
+\<<set $currentPassage.availableRelationships to lib.docksData.get.customers>>
 \<h1 class="interactive-only"><<print $building.name.toUpperFirst()>></h1>
 \<p>You walk down to the docks, which are known best for $building.notableFeature. You notice $building.notice.</p>
 \<p>The docks are $building.sizeDescriptive, and is $building.cleanlinessDescriptive. It's currently $building.activity.</p>
-<<button "Create a ship">><<run setup.docks.ships.create($town, $building)>><<update>><</button>>
+<<button "Create a ship">><<run lib.docksData.ships.create($town, $building)>><<update>><</button>>
 <<include "ShipList">>
 <<details "CreateNewNpc" "People Around">>
